Update (high- )score function
I added an info panel on the top of the screen, in it we see the current score on top left, and the highscore at the top right. The high score is saved in an external .txt file. Additionaly, we draw a line at the current highscore, so the player has a salient visual cue when they are about to pass the high score.