added enemy function and animation of enemy
I added an enemy to the game which appears once the score exceeds 1000 at random intervalls, it moves horizontally accross the screen either from left or right. 
To do that i created an extra enemy file containing the Enemy class. 
further to be able to include an animation of the enemy i added a spritesheet file with the SpriteSheet class that returns still images from the enemy png stored in the asset folder. 
enemy player collision triggers game over state. collision detection so far is based on rectangle collision. idea is for pixel collision with masks. inclusion if time allows.